flask-app/requirements.txt
Flask==2.3.2
mysql-connector-python==8.2.0
